requirements.txt
Flask==3.0.2
Flask-MySQLdb==2.0.0
mysqlclient==2.2.7
Werkzeug==3.1.3
Jinja2==3.1.6
itsdangerous==2.2.0
click==8.1.8
blinker==1.6.3
pandas==2.1.4
numpy==1.24.3
mysql-connector-python==8.2.0
openpyxl==3.1.2
xlrd==2.0.1 